chore(args)!: Remove arg field parsing to non-public
